requirements.txt
PyMySQL==0.9.3
PyJWT==1.7.1